--- packages/PLACEHOLDER-adapter/README.md
@@ -1,5 +1,0 @@
-# PLACEHOLDER adapter
-
-PLACEHOLDER adapter for salto.io
-
-TODO add details about the adapter
