+++ packages/core/package.json
@@ -55,7 +55,7 @@
     "@salto-io/workspace": "0.3.54",
     "@salto-io/zendesk-adapter": "0.3.54",
     "@salto-io/zuora-billing-adapter": "0.3.54",
-    "@salto-io/PLACEHOLDER-adapter": "0.3.54",
+    "@salto-io/serviceplaceholder-adapter": "0.3.54",
     "async-lock": "^1.2.4",
     "axios": "^0.26.1",
     "bottleneck": "^2.19.5",
