+++ packages/core/src/core/adapters/creators.ts
@@ -24,7 +24,7 @@ import { adapter as jiraAdapter } from '@salto-io/jira-adapter'
 import { adapter as stripeAdapter } from '@salto-io/stripe-adapter'
 import { adapter as oktaAdapter } from '@salto-io/okta-adapter'
 import { adapter as sapAdapter } from '@salto-io/sap-adapter'
-import { adapter as PLACEHOLDERAdapter } from '@salto-io/PLACEHOLDER-adapter'
+import { adapter as serviceplaceholderAdapter } from '@salto-io/serviceplaceholder-adapter'
 
 const adapterCreators: Record<string, Adapter> = {
   salesforce: salesforceAdapter,
@@ -39,7 +39,7 @@ const adapterCreators: Record<string, Adapter> = {
   jira: jiraAdapter,
   okta: oktaAdapter,
   dummy: dummyAdapter,
-  PLACEHOLDER: PLACEHOLDERAdapter,
+  serviceplaceholder: serviceplaceholderAdapter,
 }
 
 export default adapterCreators
