+++ packages/core/tsconfig.json
@@ -27,7 +27,7 @@
     { "path": "../workspace" },
     { "path": "../zendesk-adapter" },
     { "path": "../zuora-billing-adapter" },
-    { "path": "../PLACEHOLDER-adapter" },
+    { "path": "../serviceplaceholder-adapter" },
     { "path": "../parser" }
   ]
 }
