+++ packages/serviceplaceholder-adapter/README.md
@@ -1,0 +1,5 @@
+# serviceplaceholder adapter
+
+serviceplaceholder adapter for salto.io
+
+TODO add details about the adapter
