+++ packages/serviceplaceholder-adapter/jest.config.js
@@ -16,8 +16,8 @@
 const deepMerge = require('../../build_utils/deep_merge')
 
 module.exports = deepMerge(require('../../jest.base.config.js'), {
-  name: 'PLACEHOLDER-adapter',
-  displayName: 'PLACEHOLDER-adapter',
+  name: 'serviceplaceholder-adapter',
+  displayName: 'serviceplaceholder-adapter',
   rootDir: `${__dirname}`,
   collectCoverageFrom: ['!<rootDir>/index.ts'],
   testEnvironment: undefined,
