+++ packages/serviceplaceholder-adapter/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@salto-io/PLACEHOLDER-adapter",
+  "name": "@salto-io/serviceplaceholder-adapter",
   "version": "0.3.54",
   "license": "Apache-2.0",
-  "description": "Salto PLACEHOLDER adapter",
+  "description": "Salto serviceplaceholder adapter",
   "repository": {
     "type": "git",
     "url": "https://github.com/salto-io/salto"
