+++ packages/serviceplaceholder-adapter/src/auth.ts
