+++ packages/serviceplaceholder-adapter/src/config.ts
