+++ packages/serviceplaceholder-adapter/src/definitions/deploy/business_hours_schedule.ts
