+++ packages/serviceplaceholder-adapter/src/definitions/fetch/fetch.ts
