+++ packages/serviceplaceholder-adapter/src/definitions/fetch/index.ts
