+++ packages/serviceplaceholder-adapter/src/definitions/fetch/transforms/business_hours_schedule.ts
