+++ packages/serviceplaceholder-adapter/src/definitions/index.ts
