+++ packages/serviceplaceholder-adapter/src/definitions/references.ts
