+++ packages/serviceplaceholder-adapter/src/definitions/types.ts
