+++ packages/serviceplaceholder-adapter/test/fetch_mock_replies.json
