+++ packages/serviceplaceholder-adapter/tsconfig.json
